kserve/llm/requirements.txt
kserve==0.14.0
transformers[torch]
python-dotenv